intermediate1/requirements-dev.txt
django<2.1
djangorestframework<3.8.3